Project revenue
The `revenueByYear` object passed to the function `projectRevenue` is
ordered oldest to latest. This order posed a challenge when calculating
year-on-year averages like threeYear, fiveYear, and tenYear due to the
sequential nature of the averaging logic.

The function directly operates on `revenueByYear`, converting it into an
array (revenueAmounts) and then reversing it (reversedRevenueAmounts).
This allows for straightforward computation of year-on-year growth
rates and averages starting from the latest year and moving backwards.